# docs/swagger.yaml
definitions:
  common.ResponseHTTP:
    properties:
      details:
        type: string
      itmes: {}
      success:
        type: boolean
    type: object
  common.ResponsePagination:
    properties:
      data: {}
      details:
        type: string
      page:
        type: integer
      pages:
        type: integer
      size:
        type: integer
      success:
        type: boolean
      total:
        type: integer
    type: object
  models.Page:
    properties:
      App:
        type: string
      active:
        type: boolean
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      routes:
        items:
          $ref: '#/definitions/models.RouteResponse'
        type: array
    type: object
  models.Role:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      routes:
        items:
          $ref: '#/definitions/models.RouteResponse'
        type: array
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.RouteResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      pages:
        items:
          $ref: '#/definitions/models.Page'
        type: array
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      route_path:
        type: string
    type: object
  models.User:
    properties:
      date_registered:
        type: string
      disabled:
        type: boolean
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      uuid:
        type: string
    type: object
...
  responses.PageGet:
    properties:
      App:
        type: string
      active:
        type: boolean
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      routes:
        items:
          $ref: '#/definitions/models.RouteResponse'
        type: array
    type: object
  responses.PagePost:
    description: Page type information Contains id name and description
    properties:
      app:
        type: string
      description:
        type: string
      name:
        type: string
    required:
    - app
    - description
    - name
    type: object
...
  responses.RoutePost:
    description: Route type information Contains id name and description
    properties:
      description:
        example: Fetchs user list
        type: string
      name:
        example: get users
        type: string
      route_path:
        example: /route/path
        type: string
    required:
    - description
    - name
    - route_path
    type: object
...
info:
  contact: {}
  description: This is User Admin server.
  termsOfService: http://swagger.io/terms/
  title: Swagger Blue API
  version: "0.1"
paths:
...
  /api/pageroute/{page_id}:
    get:
      consumes:
      - application/json
      description: Get page by ID
      parameters:
      - description: Page ID
        in: path
        name: page_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/responses.PageGet'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Get Page Routes by ID
      tags:
      - Page
  /api/pageroute/{page_id}/{route_id}:
    delete:
      consumes:
      - application/json
      description: Delete Page Route
      parameters:
      - description: Page ID
        in: path
        name: page_id
        required: true
        type: integer
      - description: Route ID
        in: path
        name: route_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/responses.RoutePost'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      summary: Add Route
      tags:
      - PageRoute Operations
    post:
      consumes:
      - application/json
      description: Add Page Route
      parameters:
      - description: Page ID
        in: path
        name: page_id
        required: true
        type: integer
      - description: Route ID
        in: path
        name: route_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/responses.RoutePost'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      summary: Add Route to Page
      tags:
      - PageRoute Operations
  /api/pages:
    get:
      consumes:
      - application/json
      description: Get Pages
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: page size
        in: query
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponsePagination'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/responses.PageGet'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      - Refresh: []
      summary: Get Pages
      tags:
      - Page
    post:
      consumes:
      - application/json
      description: Add Page
      parameters:
      - description: Add Page
        in: body
        name: page
        required: true
        schema:
          $ref: '#/definitions/responses.PagePost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/responses.PagePost'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Add a new Page
      tags:
      - Page
  /api/pages/{id}:
    delete:
      consumes:
      - application/json
      description: Remove page by ID
      parameters:
      - description: Page ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Remove Page by ID
      tags:
      - Page
    get:
      consumes:
      - application/json
      description: Get page by ID
      parameters:
      - description: Page ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/responses.PageGet'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Get Page by ID
      tags:
      - Page
    patch:
      consumes:
      - application/json
      description: Patch Page
      parameters:
      - description: Patch Page
        in: body
        name: page
        required: true
        schema:
          $ref: '#/definitions/responses.PagePost'
      - description: Page ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/responses.PagePost'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Patch Page
      tags:
      - Page
...
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: X-APP-TOKEN
    type: apiKey
  Refresh:
    description: Description for what is this security definition being used
    in: header
    name: X-REFRESH-TOKEN
    type: apiKey
swagger: "2.0"